move topn heap
